+++ packages/opentelemetry-instrumentation-anthropic/tests/cassettes/test_completion/test_anthropic_completion.yaml
@@ -1,0 +1,70 @@
+interactions:
+- request:
+    body: '{"max_tokens_to_sample": 2048, "model": "claude-instant-1.2", "prompt":
+      "\n\nHuman:\nHello world\n\n\nAssistant:", "top_p": 0.1}'
+    headers:
+      accept:
+      - application/json
+      accept-encoding:
+      - gzip, deflate, br
+      anthropic-version:
+      - '2023-06-01'
+      connection:
+      - keep-alive
+      content-length:
+      - '128'
+      content-type:
+      - application/json
+      host:
+      - api.anthropic.com
+      user-agent:
+      - Anthropic/Python 0.9.0
+      x-stainless-arch:
+      - arm64
+      x-stainless-async:
+      - 'false'
+      x-stainless-lang:
+      - python
+      x-stainless-os:
+      - MacOS
+      x-stainless-package-version:
+      - 0.9.0
+      x-stainless-runtime:
+      - CPython
+      x-stainless-runtime-version:
+      - 3.9.5
+    method: POST
+    uri: https://api.anthropic.com/v1/complete
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAA4yOQQrCMBQFr6JvnYq1uxygBBfSistCiO2nBpOfahJQxLuLRcGlyzcww3sg3SeC
+        RB/85CjZwBCww5fodbmr4yHbpr5Vrdoem+pU7c9jhvhVJBaKnAtLCMQUJn0lE2c+r0iXTNwTBHwY
+        yL3rzuSBCssxGU5Fudp8VEh03LHK3rCEgAuj/ufP8wUAAP//AwCrEt2bywAAAA==
+    headers:
+      CF-Cache-Status:
+      - DYNAMIC
+      CF-RAY:
+      - 8545870d6dc009c7-HFA
+      Connection:
+      - keep-alive
+      Content-Encoding:
+      - gzip
+      Content-Type:
+      - application/json
+      Date:
+      - Mon, 12 Feb 2024 14:30:06 GMT
+      Server:
+      - cloudflare
+      Transfer-Encoding:
+      - chunked
+      Via:
+      - 1.1 google
+      X-Cloud-Trace-Context:
+      - 96aea32e8c59fb4bb3219fed5a3aac2e
+      request-id:
+      - req_01P66j1p8yrFoDi9tE7kCuA5
+    status:
+      code: 200
+      message: OK
+version: 1
