+++ packages/opentelemetry-instrumentation-openai/tests/conftest.py
@@ -1,8 +1,6 @@
 """Unit tests configuration module."""
 
 import pytest
-import yaml
-import vcr as vcr_module
 from openai import OpenAI
 from opentelemetry import trace
 from opentelemetry.sdk.trace import TracerProvider
@@ -11,14 +9,6 @@ from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.instrumentation.openai import OpenAIInstrumentor
 
 pytest_plugins = []
-
-
-class YamlSerializer:
-    def deserialize(cassette_string):
-        return yaml.load(cassette_string, encoding=("utf-8"))
-
-    def serialize(cassette_dict):
-        return yaml.dump(cassette_dict, encoding=("utf-8"))
 
 
 @pytest.fixture(scope="session")
@@ -33,17 +23,6 @@ def exporter():
     OpenAIInstrumentor().instrument()
 
     return exporter
-
-
-@pytest.fixture(scope="module")
-def vcr():
-    vcr_instance = vcr_module.VCR(
-        cassette_library_dir="tests/fixtures/cassettes",
-        record_mode="once",
-        filter_headers=["authorization"],
-    )
-    vcr_instance.register_serializer("yaml-improved", YamlSerializer())
-    return vcr_instance
 
 
 @pytest.fixture(autouse=True)
