+++ packages/opentelemetry-instrumentation-openai/tests/test_chat.py
@@ -1,23 +1,20 @@
 def test_chat_streaming(exporter, openai_client, vcr):
-    with vcr.use_cassette("chat_streaming.yaml"):
-        response = openai_client.chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=[
-                {"role": "user", "content": "Tell me a joke about opentelemetry"}
-            ],
-            stream=True,
-        )
+    response = openai_client.chat.completions.create(
+        model="gpt-3.5-turbo",
+        messages=[{"role": "user", "content": "Tell me a joke about opentelemetry"}],
+        stream=True,
+    )
 
-        for part in response:
-            pass
+    for part in response:
+        pass
 
-        spans = exporter.get_finished_spans()
-        assert [span.name for span in spans] == [
-            "openai.chat",
-        ]
-        open_ai_span = spans[0]
-        assert (
-            open_ai_span.attributes["llm.prompts.0.content"]
-            == "Tell me a joke about opentelemetry"
-        )
-        assert open_ai_span.attributes.get("llm.completions.0.content")
+    spans = exporter.get_finished_spans()
+    assert [span.name for span in spans] == [
+        "openai.chat",
+    ]
+    open_ai_span = spans[0]
+    assert (
+        open_ai_span.attributes["llm.prompts.0.content"]
+        == "Tell me a joke about opentelemetry"
+    )
+    assert open_ai_span.attributes.get("llm.completions.0.content")
