--- packages/traceloop-sdk/tests/test_openai_instrumentation.py
@@ -1,169 +1,0 @@
-import json
-from traceloop.sdk.decorators import workflow, task
-
-
-def test_embeddings(exporter, openai_client):
-    openai_client.embeddings.create(
-        input="Tell me a joke about opentelemetry",
-        model="text-embedding-ada-002",
-    )
-
-    spans = exporter.get_finished_spans()
-    assert [span.name for span in spans] == [
-        "openai.embeddings",
-    ]
-    open_ai_span = spans[0]
-    assert (
-        open_ai_span.attributes["llm.prompts.0.content"]
-        == "Tell me a joke about opentelemetry"
-    )
-    assert open_ai_span.attributes["llm.request.model"] == "text-embedding-ada-002"
-    assert open_ai_span.attributes["llm.usage.prompt_tokens"] == 8
-
-
-def test_simple_workflow(exporter, openai_client):
-    @task(name="joke_creation")
-    def create_joke():
-        completion = openai_client.chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=[
-                {"role": "user", "content": "Tell me a joke about opentelemetry"}
-            ],
-        )
-        return completion.choices[0].message.content
-
-    @workflow(name="pirate_joke_generator")
-    def joke_workflow():
-        create_joke()
-
-    joke_workflow()
-
-    spans = exporter.get_finished_spans()
-    assert [span.name for span in spans] == [
-        "openai.chat",
-        "joke_creation.task",
-        "pirate_joke_generator.workflow",
-    ]
-    open_ai_span = spans[0]
-    assert (
-        open_ai_span.attributes["llm.prompts.0.content"]
-        == "Tell me a joke about opentelemetry"
-    )
-    assert open_ai_span.attributes.get("llm.completions.0.content")
-
-
-def test_chat_streaming(exporter, openai_client):
-    response = openai_client.chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=[{"role": "user", "content": "Tell me a joke about opentelemetry"}],
-        stream=True,
-    )
-
-    for part in response:
-        pass
-
-    spans = exporter.get_finished_spans()
-    assert [span.name for span in spans] == [
-        "openai.chat",
-    ]
-    open_ai_span = spans[0]
-    assert (
-        open_ai_span.attributes["llm.prompts.0.content"]
-        == "Tell me a joke about opentelemetry"
-    )
-    assert open_ai_span.attributes.get("llm.completions.0.content")
-
-
-def test_vision(exporter, openai_client):
-    response = openai_client.chat.completions.create(
-        model="gpt-4-vision-preview",
-        messages=[
-            {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": "What is in this image?"},
-                    {
-                        "type": "image_url",
-                        "image_url": {
-                            "url": "https://source.unsplash.com/8xznAGy4HcY/800x400"
-                        },
-                    },
-                ],
-            }
-        ],
-    )
-
-    for part in response:
-        pass
-
-    spans = exporter.get_finished_spans()
-    assert [span.name for span in spans] == [
-        "openai.chat",
-    ]
-    open_ai_span = spans[0]
-    assert json.loads(open_ai_span.attributes["llm.prompts.0.content"]) == [
-        {"type": "text", "text": "What is in this image?"},
-        {
-            "type": "image_url",
-            "image_url": {"url": "https://source.unsplash.com/8xznAGy4HcY/800x400"},
-        },
-    ]
-
-    assert open_ai_span.attributes.get("llm.completions.0.content")
-
-
-def test_open_ai_function_calls(exporter, openai_client):
-    @task(name="function_call_test")
-    def function_call_test():
-        completion = openai_client.chat.completions.create(
-            model="gpt-4",
-            messages=[
-                {"role": "user", "content": "What's the weather like in Boston?"}
-            ],
-            functions=[
-                {
-                    "name": "get_current_weather",
-                    "description": "Get the current weather in a given location",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The city and state, e.g. San Francisco, CA",
-                            },
-                            "unit": {
-                                "type": "string",
-                                "enum": ["celsius", "fahrenheit"],
-                            },
-                        },
-                        "required": ["location"],
-                    },
-                }
-            ],
-            function_call="auto",
-        )
-        return completion.choices[0].message.content
-
-    @workflow(name="function_call_test_workflow")
-    def function_call_test_workflow():
-        function_call_test()
-
-    function_call_test_workflow()
-
-    spans = exporter.get_finished_spans()
-    open_ai_span = spans[0]
-    assert (
-        open_ai_span.attributes["llm.prompts.0.content"]
-        == "What's the weather like in Boston?"
-    )
-    assert (
-        open_ai_span.attributes["llm.request.functions.0.name"] == "get_current_weather"
-    )
-    assert (
-        open_ai_span.attributes["llm.request.functions.0.description"]
-        == "Get the current weather in a given location"
-    )
-    assert (
-        open_ai_span.attributes["llm.completions.0.function_call.name"]
-        == "get_current_weather"
-    )
